Got Preston's question recylcer working with my code
