+++ app/src/main/res/layout/question_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".question_view">
+    tools:context=".QuestionActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -80,7 +80,6 @@
             android:layout_marginStart="10dp"
             android:layout_marginEnd="10dp"
             />
-
 
     </LinearLayout>
